move collectionspage component
